requirements.txt
Django==1.8.3
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.20
celery==3.1.18
django-celery==3.1.16
djangorestframework==3.1.3
kombu==3.0.26
psycopg2==2.6.1
pypdfocr==0.8.2
python-dateutil==2.4.2
python-mimeparse==0.1.4
pytz==2015.4
six==1.9.0
